requirements.txt
Flask==3.0.0
Werkzeug==3.0.1

pdfplumber==0.11.0
python-docx==1.1.0
PyMuPDF==1.24.10   # <--- add this line

gunicorn==21.2.0
python-dotenv==1.0.1